Rebuilding
styled the elements in html then had a problem with the button element. i used a code that i already tried and it worked but somehow when i tried the same code in this project, it was behaving differently. after troubleshooting for hours, i came to the conclusion that it's not the code that is the problem but there's something else that until now, i don't know what is the problem but i solved it through asking chatgpt different css styles to use so i can still achieve the same output without using the first code and i ended up just modifying the code for button. instead of calling button alone, i used class